Refactor user type exports for improved modularity; separate user-related types into individual files for better organization and maintainability. Update test imports to reflect new structure.
