+++ backend/src/apps/user/types/CreateUserRequest.ts
@@ -1,0 +1,10 @@
+export interface CreateUserRequest {
+  email: string;
+  password: string;
+  username: string;
+  firstName?: string;
+  lastName?: string;
+  profileName?: string;
+  profileBio?: string;
+  profilePicture?: string; // Cloudinary URL
+}
