+++ backend/src/apps/user/types/DatabaseUser.ts
@@ -1,0 +1,6 @@
+import { User } from './User';
+
+// Database user includes password for auth operations
+export interface DatabaseUser extends User {
+  password: string;
+}
