+++ backend/src/apps/user/types/UserSearchResult.ts
@@ -1,0 +1,8 @@
+export interface UserSearchResult {
+  id: string;
+  username: string;
+  profileName?: string;
+  profilePicture?: string; // Cloudinary URL
+  profileBio?: string;
+  followersCount?: number;
+}
