+++ backend/src/apps/user/types/UserWithoutPassword.ts
@@ -1,0 +1,3 @@
+import { User } from './User';
+
+export interface UserWithoutPassword extends Omit<User, 'password'> {}
